+++ primitives/src/bridges/chain_circuit/mod.rs
@@ -100,7 +100,7 @@ pub use time_units::*;
 pub mod time_units {
     use super::BlockNumber;
 
-    pub const MILLISECS_PER_BLOCK: u64 = 6000;
+    pub const MILLISECS_PER_BLOCK: u64 = 12000;
     pub const SLOT_DURATION: u64 = MILLISECS_PER_BLOCK;
 
     pub const MINUTES: BlockNumber = 60_000 / (MILLISECS_PER_BLOCK as BlockNumber);