+++ runtime/common-types/src/lib.rs
@@ -44,7 +44,7 @@ pub const CONTRACTS_DEBUG_OUTPUT: bool = true;
 /// up by `pallet_aura` to implement `fn slot_duration()`.
 ///
 /// Change this to adjust the block time.
-pub const MILLISECS_PER_BLOCK: u64 = 6000;
+pub const MILLISECS_PER_BLOCK: u64 = 12_000;
 
 // NOTE: Currently it is not possible to change the slot duration after the chain has started.
 //       Attempting to do so will brick block production.
@@ -85,3 +85,9 @@ parameter_types! {
 pub type Address = sp_runtime::MultiAddress<AccountId, ()>;
 /// Block header type as expected by this runtime.
 pub type Header = generic::Header<BlockNumber, BlakeTwo256>;
+
+#[test]
+fn fixed_block_time_12s() {
+    assert_eq!(MILLISECS_PER_BLOCK, 12_000);
+    assert_eq!(SLOT_DURATION, 12_000);
+}